+++ webapp/app/scripts/malmsturmgm/locale.js
@@ -17,7 +17,14 @@ app.config(['$translateProvider',function ($translateProvider) {
       CHARACTER: 'Charakter',
       CHARACTERSELECTION: 'Wähle die Charakter aus',
       PDFPREF: 'PDF Einstellungen',
-      'PRINTASPEKTTEXT': 'Text der Aspekte mit ausgeben'
+      PRINTASPEKTTEXT: 'Text der Aspekte mit ausgeben',
+      DEFAULT_EMPTY_CONFIG_BLOCK:
+'{\n\
+  "docId": "",\n\
+  "file_version": 1,\n\
+  "characters": [],\n\
+  "pdftemplate": "templateMalmsturmGM1"\n\
+}'
     }
   })
 }]);