+++ webapp/app/scripts/openrole/DialogOpenCtrl.js
@@ -1,13 +1,16 @@
 "use strict";
 
-var DialogOpenCtrl = ['$scope', '$modalInstance', '$http', '$rootScope', 'characterlist', function ($scope, $modalInstance, $http, $rootScope, characterlist) {
+var DialogOpenCtrl = ['$scope', '$modalInstance', '$http', '$rootScope','characterlist','titleResKey','okActionRef',
+  function ($scope, $modalInstance, $http, $rootScope, characterlist, titleResKey, okActionRef) {
     $scope.characterlist = characterlist;
+    $scope.titleResKey = titleResKey;
+    $scope.okAction = okActionRef;
     $scope.selected = {
         char: $scope.characterlist[0]
     };
-    $scope.openCharacter = function () {
+    $scope.open = function () {
         console.log("open char "+$scope.selected.char.charactername);
-        $modalInstance.close($scope.selected.char.docId); // parameter is the character json
+        $modalInstance.close({'docId':$scope.selected.char.docId, 'action':$scope.okAction}); // parameter is the character json
     };
 
     $scope.cancel = function () {