+++ webapp/app/scripts/openrole/locale.js
@@ -49,6 +49,7 @@ app.config(['$translateProvider', function ($translateProvider) {
     GENERATING_PDF : 'Generiere PDF',
     ERROR_LOAD_IMG : 'Konnte Bilder nicht laden. Grund:',
     OPEN_CHARACTER_TITLE: 'Öffne einen Charakter',
+    OPEN_WITH_CUSTOM_CONF_TITLE: 'Konfiguration auswählen',
     PDFTEMPLATE:'PDF-Template',
     MENU: {
       LOGOUT: 'Abmelden',
@@ -81,7 +82,7 @@ app.config(['$translateProvider', function ($translateProvider) {
       DATAPROTECTION: 'Die E-Mail und das Passwort werden unumkehrbar verschlüsselt (SHA1 Hash) abgespeichert und können daher nicht wiederhergestellt werden. Nur zum Login und für die Passwort vergessen Funktion wird die E-Mail übertragen.\n'+
         'Die E-Mail wird nie geprüft, nicht weitergegeben und wird in keinster Weise weiterverwendet außer für die Passwort vergessen Funktion.\n'+
         'Möchtest du auf die Passwort vergessen Funktion verzichten, kannst du einen beliebigen Usernamen benutzen, das hat zur Folge, dass bei einem Verlust des Passworts der Zugang nur mit einer freiwilligen Spende und einer E-Mail über das Support-System wiederhergestellt werden.',
-      SHARES: 'Lese-Freigaben für folgende Spielleiter'
+      SHARES: 'Lese-Freigaben für folgende Benutzer (Nick)'
     },
     BUTTON: {
       OPEN: 'Öffnen',